bye pragma once.
In the C and C++ programming languages, #pragma once is a non-standard but widely supported preprocessor directive designed to cause the current source file to be included only once in a single compilation.  #pragma once serves the same purpose as include guards, but with several advantages, including: less code, avoidance of name clashes / duplication, and sometimes improvement in compilation speed.