ResourceScript functionality separation:
Moving .dll to the right library folder after compiling.

When creating a ScriptComponent, first search for a ResourceScript with the same ScriptName. If you find it, just reference. If not, create one with this name, reference it and store inside ModuleResources.

ResourceScript has all the assembly, image and callback methods references, but the ScriptComponent has an instance of this class, which will be different for different components referencing the same script.